+++ Cocoa/Classes/RKDocumentController.m
@@ -4,7 +4,7 @@
 
 @implementation RKDocumentController
 
-- (int)runModalOpenPanel:(NSOpenPanel *)openPanel forTypes:(NSArray *)extensions
+- (NSInteger)runModalOpenPanel:(NSOpenPanel *)openPanel forTypes:(NSArray *)extensions
 {
 	// set-up open panel (this happens every time, but no harm done)
 	ApplicationDelegate *appDelegate = [NSApp delegate];
